added unlinking tokens

+++ UserAuth/views.py
@@ -80,9 +80,11 @@ def profile_view(request):
     return render(request, 'UserAuth/profile.html', {"associated_spotify_tokens": associated_spotify_tokens})
 
 
-def delete_token(request, spotify_account_username):
-    SpotifyToken.objects.filter(spotify_account_username=spotify_account_username).delete()
-    return JsonResponse({"success":True})
+def delete_token(request):
+    if request.method == "POST":
+        spotify_account_username = request.POST.get('spotify_account_username')
+        SpotifyToken.objects.filter(spotify_account_username=spotify_account_username).delete()
+        return JsonResponse({"success":True})
 
 
 def logout_view(request):
